Use the Dialog component. Separate the entire OTP dialog to a component otpinput to reuse in many other screens
